python_document_extractors/requirements.txt
# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
PyMuPDF==1.23.8
pytesseract==0.3.10
Pillow==10.1.0

# Word Document Processing
python-docx==0.8.11
docx2txt==0.8

# PowerPoint Processing
python-pptx==0.6.21

# Excel Processing
openpyxl==3.1.2
xlrd==2.0.1
pandas==2.1.4

# Text Processing
chardet==5.2.0

# General
requests==2.31.0

